Disable react/jsx-first-prop-new-line
This is not a breaking change because this does not change the warning.

Motivation
------------

This rule is enabled by me in https://github.com/voyagegroup/eslint-config-fluct/issues/33
However, this was my mistake. In a real world, we face the case of that
we'd like to take both form of these:

```javascript
// style 1
<a href={href} className={...}>{'some link'}</a>

// style 2
<a href={href}
  className={...}
>
  {'some link'}
</a>

// style 3
<a
  href={href}
  className={...}
>
  {'some link'}
</a>
```

So this change disable `react/jsx-first-prop-new-line`
as it is just a stylistic issue.