Relax to use .setState()
This restriction was my mistake.

We should prefer `.props` than `.state` in React.
But `.setState()` is a part of React and it would not a right to restrict `.setState()` to create a complex component.

I agree that we should avoid the complexity in view (React) area,
but I also think we catch them in a code review process or a design phase.